src/main/java/com/csci5308/g17/facility/Facility.java
package com.csci5308.g17.facility;

public class Facility {

    private Integer id;
    private Integer occupancy;
    private Integer managerId;
    private String name;
    private String description;
    private String location;
    private Integer timeSlotDuration;
    private Boolean active;
    private Boolean approvalRequired;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public Integer getOccupancy() {
        return occupancy;
    }

    public void setOccupancy(Integer occupancy) {
        this.occupancy = occupancy;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }

    public int getManagerId() {
        return managerId;
    }

    public void setManagerId(int managerId) {
        this.managerId = managerId;
    }

    public Integer getTimeSlot() {
        return timeSlotDuration;
    }

    public void setTimeSlot(Integer timeSlotDuration) {
        this.timeSlotDuration = timeSlotDuration;
    }

    public Boolean getActive() {
        return active;
    }

    public void setActive(Boolean active) {
        this.active = active;
    }

    public Boolean getApprovalRequired() {
        return approvalRequired;
    }

    public void setApprovalRequired(Boolean approvalRequired) {
        this.approvalRequired = approvalRequired;
    }
}
